Moved files, renamed, fix byte error
