+++ tests/assets/ecr_sample_input_bad.xml
