+++ tests/assets/ecr_sample_input_good.xml
